add component/data file, move photoMatchImg.js to data folder
